+++ assemblyadapter-list/src/androidTest/java/com/github/panpf/assemblyadapter/list/test/AssemblyListAdapterTest.kt
@@ -98,15 +98,19 @@ class AssemblyListAdapterTest {
     fun testMethodGetCount() {
         AssemblyListAdapter<Text>(listOf(TextItemFactory())).apply {
             Assert.assertEquals(0, count)
+            Assert.assertEquals(0, itemCount)
 
             submitList(listOf(Text("hello")))
             Assert.assertEquals(1, count)
+            Assert.assertEquals(1, itemCount)
 
             submitList(listOf(Text("hello"), Text("world")))
             Assert.assertEquals(2, count)
+            Assert.assertEquals(2, itemCount)
 
             submitList(null)
             Assert.assertEquals(0, count)
+            Assert.assertEquals(0, itemCount)
         }
     }
 
@@ -300,4 +304,6 @@ class AssemblyListAdapterTest {
             getItemFactoryByPosition(1)
         }
     }
+
+    // todo test hasObservers
 }