+++ assemblyadapter-list/src/androidTest/java/com/github/panpf/assemblyadapter/list/test/expandable/AssemblyExpandableListAdapterTest.kt
@@ -548,4 +548,7 @@ class AssemblyExpandableListAdapterTest {
             getItemFactoryByPosition(1)
         }
     }
+
+    // todo test hasObservers
+    // todo test isChildSelectable
 }