+++ assemblyadapter-list/src/androidTest/java/com/github/panpf/assemblyadapter/list/test/expandable/AssemblySingleDataExpandableListAdapterTest.kt
@@ -394,4 +394,7 @@ class AssemblySingleDataExpandableListAdapterTest {
             }
         }
     }
+
+    // todo test hasObservers
+    // todo test isChildSelectable
 }